Refactor hooks into separate module
