+++ hooks.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <Windows.h>
+
+DWORD WINAPI InitHook(LPVOID);
+void OpenDebugConsole();
