+++ module_utils.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <Windows.h>
+#include <cstdint>
+#include <vector>
+#include <string>
+#include "vacstructs.h"
+
+bool DecryptVacModule(uint8_t* base, size_t imgSize, const uint8_t key[8]);
+bool DumpVacModule(VacModuleInfo_t* m, const std::wstring& dumpDir = L"C:\\Lumina");
+void FixVacModule(DWORD pImage, DWORD pModule);
+size_t GetAllocationSize(void* startAddress);
